cli updates
