+++ helper/init-env.js
@@ -3,11 +3,11 @@ const path = require("path");
 
 const initEnv = async (LIBRARY_NAME) => {
   let envFileContents = fs.readFileSync(
-    path.resolve(process.cwd(), `node_modules/${LIBRARY_NAME}/env.json`),
+    path.resolve(`./node_modules/${LIBRARY_NAME}/env.json`),
     "utf8"
   );
   let projectEnvFileContents = fs.readFileSync(
-    path.resolve(process.cwd(), `src/config/config.json`),
+    path.resolve(`./src/config/config.json`),
     "utf8"
   );
   envFileContents = JSON.parse(envFileContents);
@@ -48,7 +48,7 @@ const initEnv = async (LIBRARY_NAME) => {
     });
 
     fs.writeFileSync(
-      path.resolve(process.cwd(), `src/config/config.json`),
+      path.resolve(`./src/config/config.json`),
       JSON.stringify(projectEnvFileContents, null, 2),
       "utf8"
     );
@@ -68,7 +68,7 @@ const initEnv = async (LIBRARY_NAME) => {
     }
 
     fs.writeFileSync(
-      path.resolve(process.cwd(), `src/config/config.json`),
+      path.resolve(`./src/config/config.json`),
       JSON.stringify(projectEnvFileContents, null, 2),
       "utf8"
     );
