+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@juego/njs2-cli",
-  "version": "1.0.0",
+  "name": "@njs2/cli",
+  "version": "1.4.0",
   "description": "",
   "main": "index.js",
   "bin": {
